+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - Fix Jira metadata collector crash on duplicate field names (OSIDB-4920)
+- Fixed double error raise problem in sync_manager.
 
 ## [5.11.0] - 2026-05-28
 ### Added